Feature/sprint 5

+++ Api/Contracts/SupplyChain/CreateItemRequestDTO.cs
@@ -5,7 +5,7 @@ public class CreateItemRequest
     public int UomId { get; set; }
     public int CategoryId { get; set; }
     public string ItemName { get; set; } = string.Empty;
-    public int MinStockLevel { get; set; }
-    public int MaxStockLevel { get; set; }
+    public decimal MinStockLevel { get; set; }
+    public decimal MaxStockLevel { get; set; }
     public bool IsActive { get; set; } = true;
 }
