+++ Api/Contracts/SupplyChain/UpdateItemRequestDTO.cs
@@ -5,7 +5,7 @@ public class UpdateItemRequest
     public int? UomId { get; set; }
     public int? CategoryId { get; set; }
     public string? ItemName { get; set; }
-    public int? MinStockLevel { get; set; }
-    public int? MaxStockLevel { get; set; }
+    public decimal? MinStockLevel { get; set; }
+    public decimal? MaxStockLevel { get; set; }
     public bool? IsActive { get; set; }
 }
